Fix issues with formatting date filters (#10180)
* #10170 cleanup FilterActions

* #10170 update/cleanup imports

* - code comments

* #10170 move filter actions

* #10170 filter formatting test
